apps/api-analytics/pyproject.toml
[tool.poetry]
name = "adyela-api-analytics"
version = "0.1.0"
description = "Adyela Analytics Microservice - BigQuery integration and reporting"
authors = ["Adyela Team <dev@adyela.com>"]
readme = "README.md"
packages = [{include = "adyela_api_analytics"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.9.2"
pydantic-settings = "^2.6.0"
google-cloud-bigquery = "^3.25.0"
google-cloud-pubsub = "^2.23.0"
google-cloud-firestore = "^2.18.0"
httpx = "^0.27.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
ruff = "^0.7.2"
mypy = "^1.13.0"
black = "^24.10.0"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=adyela_api_analytics --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["adyela_api_analytics"]
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.black]
line-length = 100
target-version = ["py312"]
